# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - authentication_error
    - authorization_error
    - database_error
    - forbidden
    - internal_error
    - invalid_request
    - malformed_body
    - not_implemented
    - password_too_short
    - refresh_token_invalid
    - resource_already_exists
    - resource_not_found
    - token_invalid
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceNotFound
    - ErrCodeTokenInvalid
  auth.AccessTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJTaWduYWxTZXJ2ZXIiLCJzdWIiOiI2OGZiNWY1Yi1lM2Y1LTRhOTYtOGQzNS1jZDIyMDNhMDZmNzMiLCJleHAiOjE3NDY3NzA2MzQsImlhdCI6MTc0Njc2NzAzNH0.3OdnUNgrvt1Zxs9AlLeaC9DVT6Xwc6uGvFQHb6nDfZs
        type: string
      account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      account_type:
        enum:
        - user
        - service_identity
        type: string
      expires_in:
        description: seconds
        example: 1800
        type: integer
      isn_perms:
        additionalProperties:
          type: string
        type: object
      role:
        enum:
        - owner
        - admin
        - member
        example: admin
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  database.GetIsnRetrieverByIsnSlugRow:
    properties:
      created_at:
        type: string
      default_rate_limit:
        type: integer
      isn_id:
        type: string
      isn_is_in_use:
        type: boolean
      listener_count:
        type: integer
      retriever_status:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateIsnRetrieverRequest:
    properties:
      default_rate_limit:
        description: maximum number of requests per minute per session
        example: 600
        type: integer
      isn_slug:
        example: sample-isn--example-org
        type: string
      listener_count:
        example: 1
        type: integer
    type: object
  handlers.CreateIsnRetrieverResponse:
    properties:
      resource_url:
        example: http://localhost:8080/api/isn/sample-isn--example-org/signals/retriever
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: example@example.com
        type: string
      password:
        description: passwords must be at least 11 characters long
        example: lkIB53@6O^Y
        type: string
    type: object
...
  handlers.UpdateIsnRetrieverRequest:
    properties:
      default_rate_limit:
        description: maximum number of requests per minute per session
        example: 600
        type: integer
      listener_count:
        example: 1
        type: integer
      retriever_status:
        enum:
        - offline
        - online
        - error
        - closed
        example: offline
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        example: example_error_code
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Signals ISN service API
paths:
...
  /api/isn/{isn_slug}/signals/retriever:
    get:
      parameters:
      - description: isn slug
        example: sample-isn--example-org
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.GetIsnRetrieverByIsnSlugRow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get an ISN retriever config
      tags:
      - ISN view
    post:
      description: |-
        An ISN retriever handles the http requests sent by clients to get Signals from the ISN

        You can specify how many retrievers should be started for the ISN and they will listen on an automatically generted port

        The public facing url will be hosted on https://{isn_host}/isn/{isn_slug}/signals/retriever
        the isn_host will typically be a load balancer or API gateway that proxies requests to the internal signald services

        note retrievers are created in 'offline' mode.

        This endpoint can only be used by the site owner or the ISN admin
      parameters:
      - description: isn slug
        example: sample-isn--example-org
        in: path
        name: isn_slug
        required: true
        type: string
      - description: ISN retriever details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateIsnRetrieverRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateIsnRetrieverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Create an ISN Retriever definition
      tags:
      - ISN config
    put:
      description: This endpoint can only be used by the site owner or the ISN admin
      parameters:
      - description: isn slug
        example: sample-isn--example-org
        in: path
        name: isn_slug
        required: true
        type: string
      - description: ISN retriever details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateIsnRetrieverRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Update an ISN Retriever
      tags:
      - ISN config
...
  /auth/login:
    post:
      description: |-
        The response body includes an access token. A refresh token is included in a http-only cookie named refresh_token
        The access_token is valid for 30 mins.

        Use the refresh_token with the /auth/refresh endpoint to renew the access_token.

        The refresh_token lasts 30 days unless it is revoked earlier.
        To renew the refresh_token, log in again.
      parameters:
      - description: email and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Login
      tags:
      - auth
...
produces:
- application/json
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...